ligar as telas

+++ abr3devam/src/main/java/br/com/am/entities/Rseguranca.java
@@ -2,25 +2,25 @@ package br.com.am.entities;
 
 public class Rseguranca {
 
-	private Long id_respostaSeg;
+	private int id_respostaSeg;
 	private String Rseguranca;
 
 	public Rseguranca() {
 		super();
 	}
 
-	public Rseguranca(Long id_respostaSeg, String rseguranca) {
+	public Rseguranca(int id_respostaSeg, String rseguranca) {
 		super();
 		this.id_respostaSeg = id_respostaSeg;
 		Rseguranca = rseguranca;
 
 	}
-
-	public Long getId_respostaSeg() {
+	
+	public int getId_respostaSeg() {
 		return id_respostaSeg;
 	}
 
-	public void setId_respostaSeg(Long id_respostaSeg) {
+	public void setId_respostaSeg(int id_respostaSeg) {
 		this.id_respostaSeg = id_respostaSeg;
 	}
